Ui farm (#537)
* added farm cards

* updated farming page
updated card, accordion and table components

* updated formating

* added @chia/core and @chia/icons packages
small refactoring

* replaced brand with chia icon
removed modal dialog

* removed old pages
added WalletType into the constants

* added craco for local alias
fixed typos

* fixed app provdier and typos

* eslint format

* updated deps

* added minimum K=32 for mainnet
added precision = 3 for plot size and total net size
refactored plot page
added hooks for open and select directory via electron api
added react-form-hooks components for material-ui

* updated format of the code

* fix simulator

* black

* deps

* added plot list

* removed unused components
moved plot list from farmer into the plot page
removed unused console.logs
updated format for fee, reward and total number

* Bump colorlog to 4.5.0

* Bump colorlog to 4.6.2

* Farm crash fixes (#512)

* Bump blspy to 0.2.5

Incorporate @wjblanke's potential bls fix for farming crashes

* blspy to 0.2.6

* move to blspy 0.2.7

* added tooltip into the table cell
fixed list of the plots
added farming visualisation and hook for the farming state

* renamed back index to selectKey

* transaction type

* Bump cbor2 from 5.1.2 to 5.2.0

Bumps [cbor2](https://github.com/agronholm/cbor2) from 5.1.2 to 5.2.0.
- [Release notes](https://github.com/agronholm/cbor2/releases)
- [Changelog](https://github.com/agronholm/cbor2/blob/master/docs/versionhistory.rst)
- [Commits](https://github.com/agronholm/cbor2/compare/5.1.2...5.2.0)


* Add cbor 5.2.0 to install.sh

* Don't forget azure pipelines

* Bump sortedcontainers from 2.2.2 to 2.3.0

Bumps [sortedcontainers](https://github.com/grantjenks/python-sortedcontainers) from 2.2.2 to 2.3.0.
- [Release notes](https://github.com/grantjenks/python-sortedcontainers/releases)
- [Changelog](https://github.com/grantjenks/python-sortedcontainers/blob/master/HISTORY.rst)
- [Commits](https://github.com/grantjenks/python-sortedcontainers/compare/v2.2.2...v2.3.0)


* Bump clvm-tools from 0.1.6 to 0.1.7

Bumps [clvm-tools](https://github.com/Chia-Network/clvm_tools) from 0.1.6 to 0.1.7.
- [Release notes](https://github.com/Chia-Network/clvm_tools/releases)
- [Commits](https://github.com/Chia-Network/clvm_tools/compare/0.1.6...0.1.7)


* Have Cbor2 version only rely on setup.py (#518)

* cbor2 version should only be in setup.py

* Build and install wheels in one step on Azure

* swapped standard puzzle to use aggsig_me

fixed bug in sign_coin_solutions where aggsig_me would not be signed 
correctly

* cc_wallet and trade_manager use aggsig_me

* fixed aggsig_me bug in wallet_tool

* appeased superlinter black requests

* added aggsig_me to cost_calculator, same as aggsig

* black fix cost calculator

* Bump chiapos to 0.12.35, catch up changelog

* Appease the markdown lint

* used new type enum from transaction
simplified modals and added new type of modals

* updated format of the UI files

* added failed and not found plots
fixed plot header
added delete action for plot

* typo

* fixed scroll bar in the wallets list

* removed unused code and fixed windows build

* fixed all known scroll bar issues

* added layout sidebar
simplified layout flexbox

* added radio button to chia/core
added box shadow for layout sidebar
showed plot in parallel (disabled until we add support)

* npm run buil

* removed additional fields from queue item

* first version of refactored trade part

* call get_plots after harvester
updated format
removed unused console.logs

* added plot queue

* upnp

* Bump pos/vdf/bip158/bls for python 3.9 wheels

* downgrade to last good chiavdf - no 3.9 ubuntu yet

* Build python 3.8 and 3.9 on MacOS - temporarily compile chiavdf on Mac

* Don't build vdf_client at first on Mac 3.9

* Modify typing for python 3.9 (#532)

* Give MacOS 3.9 more time to test

* try get_args

* 3.7

* sys

* tuple

* is not none

* pretty printer versus flake

* black

* Revert MacOS testing time to 60m max


* Bump concurrent-log-handler from 0.9.17 to 0.9.19 (#530)

Bumps [concurrent-log-handler](https://github.com/Preston-Landers/concurrent-log-handler) from 0.9.17 to 0.9.19.
- [Release notes](https://github.com/Preston-Landers/concurrent-log-handler/releases)
- [Changelog](https://github.com/Preston-Landers/concurrent-log-handler/blob/master/CHANGELOG.md)
- [Commits](https://github.com/Preston-Landers/concurrent-log-handler/compare/0.9.17...0.9.19)



* Bump actions/upload-artifact from v2.2.0 to v2.2.1 (#528)

Bumps [actions/upload-artifact](https://github.com/actions/upload-artifact) from v2.2.0 to v2.2.1.
- [Release notes](https://github.com/actions/upload-artifact/releases)
- [Commits](https://github.com/actions/upload-artifact/compare/v2.2.0...726a6dcd0199f578459862705eed35cda05af50b)



* Bump keyring from 21.4.0 to 21.5.0 (#529)

Bumps [keyring](https://github.com/jaraco/keyring) from 21.4.0 to 21.5.0.
- [Release notes](https://github.com/jaraco/keyring/releases)
- [Changelog](https://github.com/jaraco/keyring/blob/master/CHANGES.rst)
- [Commits](https://github.com/jaraco/keyring/compare/v21.4.0...v21.5.0)



* Try pyinstaller-4.1 (#533)

* Try aiohttp 3.7.3

* fixed mac build

* updated formating
fixed eslint
updated deps
compiled locales

* fixed locale file format

* removed unused code

* added "add plot directory" when plots are already there

* slovak translation correction

* use electron starter from dev branch

* updated router and fixed electron build

* used TXCH for testnet

* removed eslintcache from the git

* used correct translation (TXCH) for English locale

* added missing plurals library

* used same visualization for total size

* removed unused code

* update package-lock.json

* start harvester and farmer

* Bump blspy to 0.2.9 and brute force Python 3.9 timelord issue (#540)

* Is a bad chiavdf cached?

* Bump blspy

* Check the env

* rm unexpected vdf_client in python 3.9

* Remove debug echo

* Bump pos/vdf/bip158/bls for python 3.9 wheels

* downgrade to last good chiavdf - no 3.9 ubuntu yet

* Bump pos/vdf/bip158/bls for python 3.9 wheels

* downgrade to last good chiavdf - no 3.9 ubuntu yet

+++ electron-react/src/components/selectKey/SelectKey.tsx
@@ -1,31 +1,24 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Trans } from '@lingui/macro';
 import { useSelector, useDispatch } from 'react-redux';
 import styled from 'styled-components';
+import { ConfirmDialog, Flex, Button, Link, Logo } from '@chia/core';
 import {
   Card,
   Typography,
   Container,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
   Tooltip,
   List,
   ListItem,
   ListItemText,
+  ListItemSecondaryAction,
   IconButton,
 } from '@material-ui/core';
-import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
 import {
   Delete as DeleteIcon,
   Visibility as VisibilityIcon,
 } from '@material-ui/icons';
-import Button from '../button/Button';
 import LayoutHero from '../layout/LayoutHero';
-import Flex from '../flex/Flex';
-import Brand from '../brand/Brand';
 import {
   login_action,
   delete_key,
@@ -33,10 +26,10 @@ import {
   selectFingerprint,
   delete_all_keys,
 } from '../../modules/message';
-import Link from '../router/Link';
 import { resetMnemonic } from '../../modules/mnemonic';
 import type { RootState } from '../../modules/rootReducer';
 import type Fingerprint from '../../types/Fingerprint';
+import useOpenDialog from '../../hooks/useOpenDialog';
 
 const StyledFingerprintListItem = styled(ListItem)`
   padding-right: ${({ theme }) => `${theme.spacing(11)}px`};
@@ -44,45 +37,71 @@ const StyledFingerprintListItem = styled(ListItem)`
 
 export default function SelectKey() {
   const dispatch = useDispatch();
-  const [open, setOpen] = useState<boolean>(false);
+  const openDialog = useOpenDialog();
   const publicKeyFingerprints = useSelector(
     (state: RootState) => state.wallet_state.public_key_fingerprints,
   );
   const hasFingerprints =
     publicKeyFingerprints && !!publicKeyFingerprints.length;
 
-  async function handleClick(fingerprint: Fingerprint) {
+  function handleClick(fingerprint: Fingerprint) {
     dispatch(resetMnemonic());
     dispatch(selectFingerprint(fingerprint));
     dispatch(login_action(fingerprint));
-  }
-
-  function handleClickOpen() {
-    setOpen(true);
-  }
-
-  function handleClose() {
-    setOpen(false);
-  }
-
-  function handleCloseDelete() {
-    handleClose();
-    dispatch(delete_all_keys());
   }
 
   function handleShowKey(fingerprint: Fingerprint) {
     dispatch(get_private_key(fingerprint));
   }
 
-  function handleDelete(fingerprint: Fingerprint) {
-    dispatch(delete_key(fingerprint));
+  async function handleDeletePrivateKey(fingerprint: Fingerprint) {
+    const deletePrivateKey = await openDialog((
+      <ConfirmDialog
+        title={<Trans id="DeleteKey.title">Delete key</Trans>}
+        confirmTitle={<Trans id="DeleteKey.delete">Delete</Trans>}
+        cancelTitle={<Trans id="DeleteKey.back">Back</Trans>}
+        confirmColor="default"
+      >
+        <Trans id="DeleteKey.description">
+          Deleting the key will permanently remove the key from your computer,
+          make sure you have backups. Are you sure you want to continue?
+        </Trans>
+      </ConfirmDialog>
+    ));
+
+    // @ts-ignore
+    if (deletePrivateKey) {
+      dispatch(delete_key(fingerprint));
+    }
+  }
+
+  async function handleDeleteAllKeys() {
+    const deleteAllKeys = await openDialog((
+      <ConfirmDialog
+        title={<Trans id="DeleteAllKeys.title">Delete all keys</Trans>}
+        confirmTitle={<Trans id="DeleteAllKeys.delete">Delete</Trans>}
+        cancelTitle={<Trans id="DeleteAllKeys.back">Back</Trans>}
+        confirmColor="default"
+      >
+        <Trans id="DeleteAllKeys.description">
+          Deleting all keys will permanatly remove the keys from your
+          computer, make sure you have backups. Are you sure you want to
+          continue?
+        </Trans>
+      </ConfirmDialog>
+    ));
+
+    // @ts-ignore
+    if (deleteAllKeys) {
+      dispatch(delete_all_keys());
+    }
   }
 
   return (
     <LayoutHero>
       <Container maxWidth="xs">
         <Flex flexDirection="column" alignItems="center" gap={3}>
-          <Brand />
+          <Logo width={130} />
           {hasFingerprints ? (
             <Typography variant="h5" component="h1" gutterBottom>
               <Trans id="SelectKey.title">Select Key</Trans>
@@ -141,7 +160,7 @@ export default function SelectKey() {
                           <IconButton
                             edge="end"
                             aria-label="delete"
-                            onClick={() => handleDelete(fingerprint)}
+                            onClick={() => handleDeletePrivateKey(fingerprint)}
                           >
                             <DeleteIcon />
                           </IconButton>
@@ -173,7 +192,7 @@ export default function SelectKey() {
               </Button>
             </Link>
             <Button
-              onClick={handleClickOpen}
+              onClick={handleDeleteAllKeys}
               type="submit"
               variant="contained"
               color="danger"
@@ -185,33 +204,6 @@ export default function SelectKey() {
           </Flex>
         </Flex>
       </Container>
-      <Dialog
-        open={open}
-        onClose={handleClose}
-        aria-labelledby="alert-dialog-title"
-        aria-describedby="alert-dialog-description"
-      >
-        <DialogTitle id="alert-dialog-title">
-          <Trans id="DeleteAllKeys.title">Delete all keys</Trans>
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText id="alert-dialog-description">
-            <Trans id="DeleteAllKeys.description">
-              Deleting all keys will permanatly remove the keys from your
-              computer, make sure you have backups. Are you sure you want to
-              continue?
-            </Trans>
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleClose} color="secondary" autoFocus>
-            <Trans id="DeleteAllKeys.back">Back</Trans>
-          </Button>
-          <Button onClick={handleCloseDelete} color="secondary">
-            <Trans id="DeleteAllKeys.delete">Delete</Trans>
-          </Button>
-        </DialogActions>
-      </Dialog>
     </LayoutHero>
   );
 }
